cluster: increment redpanda_cluster_partitions per partition on topic creation
Increment redpanda_cluster_partitions count per `partition` when creating a topic.
It was incremented per `replica` but it was against the metric's spec.

It's decremented per `partition` on a topic deletion and incremented per
`partition` on adding partitions as well, which align with the spec.
Hence, this eventually causes a wrong total partition count being
reported via the metric.