e2e_shadow_indexing_test: limit read buffers
Multiple OOMs have been reported in the test that report a large number
of 128K spans having been allocated. It appears, from failed
allocations, to be caused by the failure to allocate buffers on the read
path, which are tuned by storage configs.

This commit updates these configs to reduce the amount of allocation in
this test, given the point of the test is to stress the number of
segment readers, not memory consumed or performance under load.

Fixes #9375