rptest: add cloud storage timing stress test
This commit introduces a generic cloud storage test:
`CloudStorageTimingStressTest`. This test uses a cloud storage workload
that enables all operations on the cloud storage log (reads, writes,
truncation from retention, compacted-reuploads, segment merges) and
periodically performs a series of configurable checks (while the
workload is on-going). All timers that drive the storage layer are
configured with a very low period, which creates "artificial load" and
increases the likelihood of races.

There are two flavours of this test. `cloud_storage_test` is the
baseline. It runs the workload and performs the checks without inserting
any extra operations. `cloud_storage_test_with_partition_moves` runs the
same workload and uses the same checks, but includes random partition
moves. In the future, we should add more operation types to this test
and make their ordering deterministic.