c/config_manager: fixed updating seen version
When configuration is being applied we gate updates with a
`_last_seen_version` of a configuration frontend. Previously the version
was updated after the configuration update was applied to the
`configration_manager` leading to a situation in which the configuration
was overridden by the previous update.

Signed-off-by: Michal Maslanka <michal@redpanda.com>
(cherry picked from commit 7771db990bf8b6fe8abc37e770f6c905eda82e7b)