tests: update consumer config in test_si_cache_space_leak
Since strict cache size enforcement was added, this test
tends to bump against the cache size limit and end up
waiting for the 5s cache trim throttle interval.

Because it used a consumer message count of 1000, that
could result in a health consumer still failing to complete
wait() in the default timeout.

We may simply use a smaller message count: this does the same
amount of I/O because the consumer loops until asked to stop,
irrespective of the message count.

Fixes https://github.com/redpanda-data/redpanda/issues/9423