cloud_storage: Add sname format to lw segment meta
The lw segment meta object can be used to create a full segment meta. It
uses the size_bytes field to derive the sname format version, where if
the size_bytes is non zero the version is 2. As we add higher versions
this derivation is not possible, so an explicit sname format field is
added to the lw segment meta.

When a redpanda broker is updated from a version which does not contain
the version in lw meta to a version which does contain this field, the
version will be derived from size_bytes as before. Once we start
setting higher versions explicitly in code, this field will reflect the
actual version.