storage: add early out for housekeeping fiber
The abort check in disk_log::gc effectively ran inline with
the housekeeping loop in log_manager. The log manager should
have be returning early on shutdown in any case, so we can
remove the check in disk_log::gc. In either case, strategic
checks at lower levels of storage remain in place.

Signed-off-by: Noah Watkins <noahwatkins@gmail.com>