txn: add correcting get_tx
When a leader of the txn coordinator encounters a txn initiated by a
previous leader it should try fetching state of the previous leader
to have as much info as possible and deal with the situation when the
state isn't available eg by terminating the session.

Introducing a single method to do it instead of implementing it ad
hoc for every method of the coordinator.