txn: make try_abort to use get_tx
Fix violation of atomicity of the consume-transform-produce loop.
The problem is caused by the automatic abort of a tx on its expiration
(try_abort) and simultaneous retry of the commit from the client. When
a commit suspects that the tx state isn't up-to-date it does its best
to fetch state from the previous leader. try_abort was operating on the
state at hand and this inconsistency causes the trouble.

fix https://github.com/redpanda-data/redpanda/issues/9571