cloud_storage: Change object upload to use iobuf
Object upload in remote accepts a string and converts it to iobuf. This
change makes the call accept an iobuf directly. This allows the upload
object function to be used for other purposes like uploading a segment
index, without an extra conversion step.

Additionally the tags for the object are made configurable, and a log
label can optionally be passed for more descriptive logging.