storage: don't compact below the log start offset
Compaction and prefix truncation (retention) of the local log are async
operations that can race, despite the fact that their futures are
ordered in `disk_log_impl::compact`. This is because `disk_log_impl::gc`
does not remove any segments inline, but simply notifies the log
eviction STM of the offset recommended by the retention policy. Segments
are removed when Raft calls into `disk_log_impl::prefix_truncate`.

Hence, compaction may be operating on segments that are being removed,
which leads to strange races.

The fix is to skip compaction of segments that are below the new start
offset which the retention policy determined.