c/members_backend: fixed calculation of unevenness error
When calculating an unevennes error we must include planned moves. When
move is planned in redpanda the allocation is only added to the replicas
that are about to be added to the replica set but not removed from the
one that are removed during partition reallocation. Fixed calculation of
unevenness error by subtracting removed replicas on the nodes they are
moved from.

Fixes: #9507

Signed-off-by: Michal Maslanka <michal@redpanda.com>
(cherry picked from commit 5880248035e12a75b8ae35a4490d29df3a12a01b)