storage: only notify eviction monitor once
Previously, the Raft layer was notified separately for each retention
policy, this meant that for topics configured with both time and size
based retention, two log snapshots and truncations could be performed
if both policies required it.

This commit reworks the application of retention policies. Each policy,
returns the offset to evict at separately. The maximum of those two
offsets is taken and the log is only truncated once in all scenarios.