archival: hanlde leader swings in housekeeping
Truncaction of the cloud log is a two step process. First, the new start
offset is determined and an STM command is replicated in order to update
the manifest start offsets. Then, async, the segments below the start
offset in the manifest are deleted.

Leadership changes can occur between these two steps. When that happens,
the first step will be repeated once a new leader is established. If the
start offset we choose to truncate at is the same, then the command to
update the start offset is replicated again. It succeeds, but produces
an error log which trips up the tests.

This commit fixes this scenario by only replicating the command to
advance the start offset if the new offset is greater than the current
one.