txn: fix segfault
when we were removing a tx from the cache we could remove a mutex
even when the current fiber was holdin it and another fiber was
waiting on it leading to the segfault.

fixing the problem by removing the mutex after leaving it and only
if nobody waits for it.

(cherry picked from commit b6bbfd6d7b70208dff71db407c847eb35d62edf5)