service_fixture: replace std::reduce with std::accumulate to fix build with GCC.
* std::reduce expects a binary op, which will be applied to the
result of dereferencing the input iterators, the results of other
binary op and init.
* std::accumulate expects a binary op, which will be applied to
init and the result of deferencing the input iterators.

The implementation of std::reduce in libstdc++[^1] requires that
the type of element, init value and return value should be the
same(or convertible, roughly speaking), which differs from that
of libc++[^2] and apparently cannot be satisfied in our case.

Here we simply replace std::reduce with std::accumulate.

[^1]: https://github.com/gcc-mirror/gcc/blob/releases/gcc-12/libstdc%2B%2B-v3/include/std/numeric#L284
[^2]: https://github.com/llvm/llvm-project/blob/release/16.x/libcxx/include/__numeric/reduce.h#L24

Signed-off-by: Jianyong Chen <baluschch@gmail.com>