storage: report reclaimed tombstone segment bytes
When a segment is closed and fully removed collect the size of the
persistent state reclaimed and report this. The addition of stat()
should have minimal minimal impact on performance since inode will
already need to be available for the actual removal which is now also
performend in parallel across all categories of persistent state.

In addition, conslidated the trivial remove_tombstone() method into
close(), its own caller.

Signed-off-by: Noah Watkins <noahwatkins@gmail.com>