ducktape: Refactor test_random_node_operations
This is a prep change for running a long running version of this test.
Running it for longer requires using a different verifier that does not
buffer final output in memory for validation. So the idea is to switch
to KgoVerifier utilities that are more scalable than the stock verifier.
We cannot just swap the verifier because KgoVerifier does not support
compaction and we want test coverage for compacted topics.

This change refactors the test class to be able to plugin a workload.

- Moves test driver and core logic into RandomNodeOperationsTestBase
- Adds start_workload and end_workload methods that child classes can
  implement to start/stop a custom workload.
- Reimplements the current test using the above pattern with default
  e2e verifier.