archival: lock segments before opening readers
A read lock is placed on any segment that the cloud storage upload loop
selects: compacted or not. The upload loop needs to convert the begin
and end offsets into the segments to file offsets for partial uploads.

Previously, in the case of compacted segment reuploads, the locks were
taken after the conversion to file offsets. When we couple this with the
fact that compacted re-uploads can occur below the latest disk log start
offset, it meant that the segment could be removed with an outstanding
reader before the locks were acquired. This bubbled up and stopped the
upload loop with "Bad file descriptor" errors.

This commit fixes the issue by taking the locks before any readers are
created. Note that the non-compacted upload path does the same thing.