archival: fix self-concatenated segment reupload
NOTE: I had to change the test_do_not_reupload_self_concatenated test
to use `archival::segment_name` instead of `segment_name`.

The compacted segment re-upload logic had a check that meant to ensure
that the selected segments are above the start offset of the local log (
in `segment_collector::find_next_segment`). If the start offset provided
to `segment_collector::find_next_segment` falls below the start of the
log, then the the first segment in the local log is selected and the
search will continue from the offset after the commited offset of the
first local segment. However, if said offest is still below the start of
the local log, the same logic is applied and we end up selecting the
same segment multiple times.

This bug in the segment selection, leads to re-uploads that that repeat
the batches in a segment. For instance, if the first segment lies
between [100, 200] and the start offset of the log is 202, we will have
repeatedly selected the first segment until the size limit of the
re-upload stopped us. An example of such a scenario can be seen in the
added test.

In order to fix this, we check advance the start offset to the start
offset of the local log at the beginning of the iteration, rather than
on every step.

(cherry picked from commit 6a233aba5ec08b1f47da155cf7a86d4ec3ade8be)