cloud_storage: Prevent re-adding a segment
Throw exception if the segment with the same parameters is added to the
manifest second time. In this case the 'replaced' and 'segments' lists
will have the same segment. The name of the segment is the same so when
the housekeeping will remove it we will lose the data completely.