tests: update upgrade test for tiered storage upgrade behavior
When we added cloud_storage_manifest_format_v2 feature flag,
this blocks manifest uploads during upgrades.  This previously
did't block purging local storage (because we could still
write archival_metadata_stm), but this was incorrect: we should
not drop local data until we have written data _and_ metadata
remotely.

Now that this is fixed, the test requires updating to reflect
that on this version transition, uploads will not happen while
the cluster is in a mixed version state.