cloud_storage: Fix fuzz test failures
The fuzz tests are failing because when the data is generated the
reactor gets stalled. When this happens the lowres_clock counter is not
updated and becomes stale (the time delta dpeneds on the duration of the
reactor stall). When the test code tries to fetch the manifest from S3
as a next step it uses cloud_storage::remote::download_manifest. This
method accepts a timestamp which is used as a deadline for the
operation. In case of failure the deadline is computed incorrectly and
operation fails by timeout.

This commit fixes the problem by invoking lowres_clock::update before
the test starts doing anything.