archival: remove timeouts from archival_metadata_stm
At a high level, the archival loop is implemented with the following
steps:

```
while not aborted:
  wait until node is leader
  sync that offsets from previous terms are applied
  while node is leader:
    upload segments based on what is in the manifest
    replicate changes to manifest via archival stm
    wait for archival stm update to be applied (updates manifest)
```

The archival loop currently expects that updates to the manifest are
either driven by it (or the housekeeping service, though ignoring this
for the sake of example).

One case where this invariant is broken is the case where the
replication/apply of archival batches times out -- the actual apply of
the op may proceed as a part of the raft::state_machine background
fiber, and the archival loop itself may proceed without waiting, meaning
the updates to the manifest may race with a subsequent iteration of the
archival loop.

This commit remediates this by removing the timeout from the
archival_metadata_stm::do_replicate_commands() call, forcing the caller
to wait indefinitely for the op to complete.