archival: avoid replicating cmd if cleanup is noop
Previously, 'ntp_archiver::garbage_collect' replicated a
'cleanup_metadata_cmd' even when no segments were removed. The manifest
was also re-uploaded each time.

This commit fixes the issue by checking if there are any segments to
remove before replicating the command. If that's not the case, indicate
to the housekeeping that there's no reasone to re-upload the manifest.

Some ducktape tests have also been updated to upload the manifest more
often upon segment uploads. This has been done in order to avoid
timeouts when waiting for data to be uploaded to cloud storage.