archival: add some logging around read locking
The cloud storage upload path acquires read locks for the segments that
it plans to upload. Previously, we have seen this acquisition time out
due to dead-locking, which in turn led to the upload loop behaving
strangely. With these new logs in place, we should be able to identify
a if the upload loop is stuck because it cannot get a read lock.