cloud_storage: Prevent reuploads with unexpected size
It is possible that the reuploaded segment is smaller than predicted. In
this case it's not safe to reupload it because adjacent segment merging
may try to reupload it second time. This commit blocks reupload if the
size doesn't match the result of the manifest scan.

(cherry picked from commit dc855e26dd4eef308b87fca87b806124bb573acd)