storage: report more accurate disk usage sizes
Prior to this commit estimation of disk sizes involved reporting
user-level data: for example the number of bytes written into the
segment file. This is not particularly accurate because it doesn't
take into account the physical allocation of space by the underlying
file system.

For example, a small file of a few bytes in a file system that does not
have any optimizations such as storing small file data in inode blocks
might require a minimum file allocation size of 4K. This can have a
dramatic impact on disk usage for workloads that produce a large
number of files on disk: under reporting disk usage.

To make accounting more accurate we instead query for file size from the
OS. It's important to use this information appropriately: it doesn't
necessarily reflect user-level data sizes for retention. Rather, it
allows us to make better decisions about space usage when scheduling
data removal.

The impact of querying the OS is expected to be small because the number
of queries >> modifications that affect size, so caching is feasible.
Generally only segments that are compacted and the active segment for
partition will have operations that invalidate the cache. However, these
are also the segments for which we would skip applying retention policy
to in a particular round of data reclamation.

Signed-off-by: Noah Watkins <noahwatkins@gmail.com>