storage: remove caching of persistent size
The intention behind caching the size of removed persistent state for
GC'd log segments was reporting: if we held a segment reference it still
may have been a GC'd tombstone. In this case reporting its size as 0 was
awkward.

However, this interface is now going to be primarily used for
determining GC schedules. In this new role it makes sense to report the
most up-to-date information.

Signed-off-by: Noah Watkins <noahwatkins@gmail.com>