storage: report reclaimable space for non-collectibles
Prior to this commit the GC estimate size would report that no data was
reclaimable for non-collectible partitions. However, even when retention
is disabled we may still want to delete data that has been uploaded to
cloud storage. This commit fixes that accounting by reporting data that
is collectible, even if it is not collectible because of normal
retention policy.

Signed-off-by: Noah Watkins <noahwatkins@gmail.com>