storage: add a generic disk usage reporting method
This commit renames `estimate_reclaim_size` method to be a more
generically named `disk_usage` method that will return information about
both estimated reclaim size as before, as well as on-disk usage
information for the entire partition.

The two are combined because reporting all of the information together
makes for a more efficient implementation. Further, the two pieces of
information will be used together as part of disk space management.

Signed-off-by: Noah Watkins <noahwatkins@gmail.com>