cloud_storage: Add failure injection to cloud storage API
Add failure injection mechanism into the cloud_storage::remote. The
mechanism can be enabled using the RP_INJECT_CLOUD_STORAGE_API_ERRORS
environment variable. The format of the env var follows:

	<frequency>:<failure-type1>:<failure-type2> ..

The value starts with frequency parameter. It's an integer greater than
0. Value 1 means that every cloud storage api call triggers a failure,
3 means every third, 10 means 1 call out of 10 etc.

The 'frequency' is followed by a list of failure types. There are four
types of failures.
  - throttle
  - no_such_key
  - failure
  - none
The 'throttle' triggers 'SlowDown' responses. The 'no_such_key' the
retryable responses, and 'failure' triggers not retryable responses.
The 'none' type doesn't trigger an error response but still generates
a sleep.

Example:

	"100:failiure:throttle:none"

The failure types can be repeated to make one of them more frequent like
this:

	"100:throttle:throttle:failure none"

The failure is implemented in the following way. First, the RNG is used
to decide if the failure has to be injected. If RNG() % Frequency is
equal to 0 then the failure is injected. Then the RNG is used to select
a failure type. Then the same approach is used to select a timeout
value. The upper limit for timeout is a timeout of the current api call.

When the generated failure is used it invokes 'ss::sleep_abortable' and
then returns a selected error. This is done in 'generate_get_failure' and
'generate_put_failure' methods. The method may return nullopt if the
failure type is 'none' but it will still invoke 'ss::sleep_abortable'.
This is needed to simulate situation when the actual failure is not
happening but the operation times out anyway.

The 'no_such_key' error can only be triggered by the 'get' request.