cluster: Add support for prefix truncation
- This commit modifies the log_eviction_stm to be a proper stm that will
scan the underlying log for new special batches of type
'prefix_truncate'.

- These new special batches are indicators to modify the start offset.
Since this is replicated to peers, eventually all peers will modify
their respective start offsets as well.

- There is a new background processing fiber added to this class as well
that attempts to write a raft snapshot , reaping as much data as
possible as close to the current start offset.

- The class is a subclass of persisted_stm and persists to the kvstore
its last applied offset and the current start offset. In the case of
failures these pieces of information can be used to startup from the
last processed offset.