cloud_storage: fix byte count on double truncates
Previously, when advancing the start offset
(`partition_manifest::advance_start_offset`), we would always subtract
the size of all present segments which are below the new start offset
from the total cloud log size. This works correctly only if there is a
`cleanup_metadata` command between every two consecutive `truncate`
commands. Otherwise, the segments in the manifest which are below the
start offset remain the same and we will subtract the ones from the
first `truncate` twice.

This scenario can occur when a leadership transfer happens
after houskeeping replicates the `truncate` command, but before
it gets to replicate the `cleanup_metadata` command. When the next
leader runs housekeeping, it might choose to advance the start offset
even further.

The fix is to simply take the current start offset into account and
start subtracting from the segment containing the current start offset.