schema_registry: Register a schema with a provided schema id
Providing a schema id allows migration from another Schema Registry,
and should only be allowed during import. The Schema Registry does
not yet support setting the mode (`PUT /mode`, `PUT /mode/subject`),
not does it support the `IMPORT` mode, so that is not checked.

If an id for a schema is provided, then the schema must either not
already exist, or the id must match the existing one.

If an id is not provided, the existing behaviour is retained; if
the schema does not already exist, an id is allocated for it,
otherwise the existing id is returned.

Signed-off-by: Ben Pope <ben@redpanda.com>
(cherry picked from commit 2821c716aecac9093d3f260a13fc860dae13bd76)