utils/human: decimal seconds formatter
{fmt} can format durations either as integer+unit, or as h:m:s. The latter
form is not convenient for technical durations. Integer number of micro
or nanoseconds is also hard to read when the duration is seconds or
more. The introduced formatter uses integer arithmetic to represent
duration as a number of seconds with decimal part as large as
the representation of the duration type.