cluster: rpc for internode quota balancer
Quota balancer types define the message to request a quota lease
the response to it.

Quota balancer service provides framework to handle quota requests.

Quota balancer frontend is to post the quota requests outside of cluster/.

The throughput_bps named type is used to communicate throughput via rpc,
and the existing kafka/quota_manager will be changed to use it.