cloud_storage: correct size tracking on recovery
Previously, when the state of archival metadata STM was recovered from a
snapshot, the total size of the cloud log was computed as the sum of all
segments which are not in the replaced list. This is incorrect if the
last command included in the snapshot is a `truncate`, as it causes
segments that are not accesible by fetch requests to be included in the
total size.

In order to fix this, the start offset is taken into account when
computing the size of the cloud log upon hydration from snapshots.

(cherry picked from commit 3ea72defb1674d933531152e4ffcf9bb6d310183)