cloud_storage: Use 2-random strategy for borrows
Use randomized strategy to pick a shard to borrow from. This is needed
to reduce performance cost of the borrow operation when the number of
shards is large. Another change limits number of borrowed clients by a
shard. The shard is only allowed to borrow `_capacity` clients so max
number of clients that it might be able to use is limited by
`_capacity*2`.