archival: update delta_offset_end to report next delta
The field is currently not expressive enough to report the HWM, the
Kafka offset just past the end of a partition. Currently each segment
records delta_offset_end as the difference between Kafka and RP offset
at the end of the log:

c:     config batch
d:     data batch
delta: computed by counting the number of non-data batches prior to the given
       offset

kafka = model - delta

 type:   c     c     d     d     d     c     c         c     c
      +-----------------------------------------+   +-----------+
      |     |     |     |     |     |     |     |   |     |     |
model:|  0     1     2    ...   138   139   140 |   | 141   142 |
      |     |     |     |     |     |     |     |   |     |     |
      +-----------------------------------------+   +-----------+
delta:   0     1     2    ...    2     2     3         4     5     6
      +-----------------------------------------+   +-----------+
      |                 |     |     |           |   |           |
kafka:   0     0     0    1..   136   137   137 |   | 137   137 |
      |                 |     |     |           |   |           |
      +-----------------------------------------+   +-----------+

In the above example, consider when the first segment ([0, 140]) is uploaded,
and the second ([141, 142]) isn't. A read replica would previously report the
HWM as 138 (140 - 3 + 1), when really it should be 137, since the highest data
batch is 136. To translate the correct HWM, we would need to translate 141 and
record its delta as 4, giving us enough information in metadata to tell that
140 is a config batch.

To remediate this, this commit changes the existing delta_offset_end field to
translate the next offset, and expecting the next Kafka offset to be translated
with the next offset.

Regarding backwards compatibility, conditionally increasing the
delta_offset_end value means older data will over-report by 1. This should be
safe, given the value is currently only used for optimizations that skip
segments with no data batches.

Fixes #9513