utils: Pop buffers in case of early detach
When there are multiple consumers in a fanout with a deep queue and some
consumers have progressed ahead, if a slow consumer exits early by
detaching, this causes many buffers in the front of the queue to have
their bitmasks set. When a fast consumer then does the next read, it
finds a buffer to read from, sets its own bit in the mask, and then
finds the buffer eligible to pop. But the invariant that the popped
buffer should be at the front of the queue is now broken.

The changes here clean up the front of the queue before setting any
masks, so that any buffers which have their mask already set to all
bits are popped out before we proceed with the current read. This
maintains the invariant.