archival: reduce mentions of LSO in archival code
"Last stable offset" is a concept used through the codebase to indicate
the offset below which ops/transactions have been fully applied. We
uphold the invariant to never upload anything at or below the LSO.

It's currently used to indicate the first offset that shouldn't be
uploaded, but the name itself doesn't convey that without knowledge of
LSO, and even then the connection between it and the upload path isn't
obvious.

This commit removes references to LSO in archival code, renaming them to
"end_offset" or "last_offset" (based on nearby code) and adding
"inclusive" and "exclusive" suffixes where appropriate.