k/fetch: introduced replica_selector
Introduced an abstraction representing replica selector. Replica
selector is only used when consumer sets the `rack` field in
`FetchRequest`. Replica selector is a generic interface but for now we
are only going to use the `select_closest_replica` implementation i.e.
selector that prefers a replica in the same rack as the consumer.

Signed-off-by: Michal Maslanka <michal@redpanda.com>