k/fetch: handle fetching from follower
When consumer sets the `rack` field in the `FetchRequest` it indicates
that it wants to consume from the closest replica. In this case current
partition leader instead of serving a read returns closest replica id
for each partition. The decision is made by the `replica_selector` by
default a broker selects replica in the same rack as the consumer.

Signed-off-by: Michal Maslanka <michal@redpanda.com>