cloud_storage: Fix race condition in the segment_collector
The segment_collector calculates the size of the segment sequence during
the initial scan and then it uses it to produce an upload candidate. The
'make_upload_candidate' method is a future with several scheduling
points. If the size of the last segment changes during the call the
method will return incorrect content length. This commit fixes this by
caching the size of the last segment.

The problem only affect adjacent segment reuploads. It shouldn't affect
compacted segment reupload because the last compacted segment is always
rolled and its size can't change.