storage: fix start_timestamp() when _start_offset==-1
_start_offset is only set once some background housekeeping
happens: on a freshly created partition with a few records
produced into it, it can be -1.  In this case when we read
the start timestamp, we should not try to find a segment
containing this negative start offset.

We may later refactor to avoid having non-empty logs carry
a confusingly -1 start_offset.