cloud_storage: Add start_kafka_offset to...
...the cloud_storage::partition_manifest class. The property overrides
start kafka offset if set. Without it the current start offset is used
(which is aligned by the segment boundary). When retention moves
start_offset the start kafka offset property gets reset. When
DeleteRecords request moves start kafka offset the start_offset is moved
accordingly to the begining of the segment that contains corresponding
kafka offset.