dt/sr: Updated serde client test
Created a SerdeClient BackgroundThreadService that will fire off the
different clients based on which client type to use.

Moved python_librdkafka_serde_client.py to the remote_scripts
directory so it can be accessed in the docker container.

Updated python_librdkafka_serde_client.py to have the same command line
arguments as the other clients.

Signed-off-by: Michael Boquard <michael@redpanda.com>