raft/recovery_stm: flush at set intervals during recovery
For large recoveries(i.e, when a node has been down for 10mins) if
flushing is delayed to the end of recovery then in order for the
follower to apply batches to it's stms it will have to read them from
disk rather than the batch cache. This can cause unneeded disk
utilization and performance instability during recovery. Therefore this
changes issues flushes to followers at fixed intervals to ensure the
follower's stms can apply batches from the cache.