config: lower `raft_transfer_leader_recovery_timeout_ms`'s default value
When leadership of a partition is explicitly transfered the consensus will
currently wait 10 seconds for the replica to recover if need be. And
during this wait period all appends to the partition are blocked. This
can cause instability during recovery when the leader balancer is trying
to move a lot leaders back to the recovered node. By reducing this
timeout to 25ms we avoid this instability and give the leader balancer
faster feedback about which replicas are still recovering.