rpk: move tuners to a dedicated struct
This will make it easier to switch between the new rpk yaml format and
the old: rather than passing around a whole old rpk configuration, we
change the tuner codebase to only require the tuner struct. We will then
add a method to convert the old struct to the new, and then pass around
only the new struct.