cloud_roles: run setenv on reactor thread
This patch changes boost tests that call setenv so that they run on a
reactor thread rather than the main thread. Without this change we end
up hitting this assertion:

  test_cloud_roles_rpunit:
  /home/nwatkins/src/redpanda/vbuild/release/clang/v_deps_build/seastar-prefix/src/seastar/src/core/memory.cc:986:
  void seastar::memory::cpu_pages::shrink(void *, size_t): Assertion
  `object_cpu_id(ptr) == cpu_id' failed.

NOTE: it is not clear that this really fixes the issue, but at least the
issue doesn't appear to happen anymore. According to this seastar-dev
thread setenv is interacting with global state and the realloc that it
performs may involve some mixture of memory accessed by different cores,
before/after the low-level allocators are swapped out by seastar.

   https://groups.google.com/g/seastar-dev/c/hSdTKn05Ei8/m/cH4bFvmEAgAJ

Overall it seems like we might want to consider getting rid of setenv
calls. There is only one in Redpanda proper (krb5 stuff).

Signed-off-by: Noah Watkins <noahwatkins@gmail.com>