partition_manifest: removed get_inserter_iterator
this is not yet implemented in segment_meta_cstore, and the method is
not used in the codebase

partition_manifest_test: test memory tracking removal

the test doesn't make sense as it was, the memory usage now is not in a
strict linear correlation with the number of segments. keeping only the
call, to test compilation and sanitizers

partition_manifest_test: boost test change

changed from REQUIRE to CHECK in a equality checker function for
segment_meta. this is useful to get all the fields that do not match.
added a final REQUIRE on the whole segment_meta, to retain the old
behaviour of stopping the test