config: reduce the default for max_kafka_throttle_delay_ms
`max_kafka_throttle_delay_ms` default value changed to 30s to align
with the default setting of `session.timeout.ms` in java client and in
librdkafka which is 45s there. By setting the default max throttle delay
lower than session timeout, it is possible to avoid heartbeat timeouts
for librdkafka clients (librdkafka does not support KIP-219 as of now
and therefore relies on broker doing throttling) with default
configuration on both client and broker.