OLV: add data_directory_inspector.py
When data_directory_inspector inspects redpanda's data directory
it looks only on the directory structure and names of the files
but doesn't check their content.

It detects files which shouldn't be there. Most likely those files
won't be an indication of garbage in data_directory but rather a sign
that data_directory_inspector got out of sync with Redpanda. That
will force us to update the inspector and keep the description of the
on-disk files in sync with Redpanda.

Another purpose of data_directory_inspector is to decode information
hidden in redpanda structure and present it in machine & human read-
able form.

For example, /raw/data/kafka/topic1/0_8/6065833-1-v1.log is repre-
sented as

  "topics": {
    "topic1": {
      "namespace": "kafka",
      "name": "topic1",
      "partitions": [
        {
          "id": 0,
          "revision": 8,
          "segments": [
            {
              "path": "/raw/data/kafka/topic1/0_8/6065833-1-v1.log",
              "start_offset": 6065833,
              "term": 1,
              "version": "v1",
              "size": 133423845
            },

How to use it?

  python3 data_directory_inspector.py --path /path/to/data/directory

`.unknown` of the output contains a list of unexpected files and dirs